Move agent-to-agent permissions to a standalone page
The cross-agent policies UI (the settings dialog's "Agents" tab) becomes
its own routed page at /agents/$slug/agent-permissions, following the
Secrets and API Logs precedent: new route + view kind in the route-state
codec, document title, header breadcrumb, and an "Agent-to-agent
connections" row in the agent home Extras card. The settings dialog drops
the tab; page content is unchanged apart from page chrome, an owner gate
(queries disabled for non-owners), and an agent_permissions_viewed event.

The home graph's "Edit permissions" affordance still targets the settings
dialog here; it is repointed at the new page in the next PR of this stack.
