+++ src/renderer/components/agents/agent-home/home-extras.tsx
@@ -19,6 +19,13 @@ export function HomeExtras({ agentSlug, onOpenSettings, className }: HomeExtrasP
     <div className={cn("rounded-xl border bg-background py-2", className)}>
       <div className="divide-y divide-border/50">
         <HomeDefaultModel agentSlug={agentSlug} />
+        <ExtrasButton
+          label="Agent-to-agent connections"
+          onClick={() => {
+            void navigate({ to: '/agents/$slug/agent-permissions', params: { slug: agentSlug } })
+          }}
+          testId="home-agent-permissions-open-page"
+        />
         <ExtrasButton label="System Prompt" onClick={() => onOpenSettings?.('system-prompt')} />
         <ExtrasButton
           label="Agent Directory"
