+++ src/renderer/components/agents/agent-settings-dialog.tsx
@@ -1,13 +1,12 @@
 
 import * as React from 'react'
-import { Settings, Users, Network } from 'lucide-react'
+import { Settings, Users } from 'lucide-react'
 import { useUser } from '@renderer/context/user-context'
 import { Button } from '@renderer/components/ui/button'
 import { SettingsDialog, SettingsDialogTab } from '@renderer/components/ui/settings-dialog'
 import { useUpdateAgent, type ApiAgent } from '@renderer/hooks/use-agents'
 import { GeneralTab } from './settings/general-tab'
 import { AccessTab } from './settings/access-tab'
-import { XAgentPoliciesTab } from './settings/x-agent-policies-tab'
 
 interface AgentSettingsDialogProps {
   agent: ApiAgent
@@ -86,9 +85,6 @@ export function AgentSettingsDialog({
           onNameChange={setName}
           onDialogClose={() => onOpenChange(false)}
         />
-      </SettingsDialogTab>
-      <SettingsDialogTab id="x-agent-policies" label="Agents" icon={<Network className="h-4 w-4" />}>
-        <XAgentPoliciesTab agentSlug={agent.slug} />
       </SettingsDialogTab>
       {isAuthMode && (
         <SettingsDialogTab id="access" label="Access" icon={<Users className="h-4 w-4" />}>
