+++ src/renderer/hooks/use-document-title.ts
@@ -105,6 +105,8 @@ function titleForAgentView(view: AgentView, input: DocumentTitleInput): string {
       return joinView(agentTitle, 'API Logs')
     case 'secrets':
       return joinView(agentTitle, 'Secrets')
+    case 'agentPermissions':
+      return joinView(agentTitle, 'Agent-to-agent Connections')
     case 'connections':
       return joinView(agentTitle, view.detail?.view === 'logs' ? 'Connection Logs' : 'Connections')
     case 'notifications':
